Moved all game definition logic into the framework itself.
