+++ CMakeLists.txt
@@ -465,6 +465,7 @@ if(COMPILATION_PLATFORM STREQUAL "WIN32")
                 "${PROJECT_SOURCE_DIR}/Source/ogrid/Player/*.cpp"
                 "${PROJECT_SOURCE_DIR}/Source/ogrid/GUI/*.cpp"
                 "${PROJECT_SOURCE_DIR}/Source/ogrid/Helpers/*.cpp"
+                "${PROJECT_SOURCE_DIR}/Source/ogrid/GameDefinition/*.cpp"
         )
 
         # OMNI-GRID .H FILES
@@ -476,6 +477,7 @@ if(COMPILATION_PLATFORM STREQUAL "WIN32")
                 "${PROJECT_SOURCE_DIR}/Source/ogrid/Player/*.h"
                 "${PROJECT_SOURCE_DIR}/Source/ogrid/GUI/*.h"
                 "${PROJECT_SOURCE_DIR}/Source/ogrid/Helpers/*.h"
+                "${PROJECT_SOURCE_DIR}/Source/ogrid/GameDefinition/*.h"
         )
 
         if(${VERBOSE})
@@ -577,15 +579,11 @@ if(COMPILATION_PLATFORM STREQUAL "WIN32")
         # ADD SOURCE FILES
         # Ogrid-Sandbox .CPP FILES
         file(GLOB_RECURSE Ogrid-Sandbox_CPP
-                "${PROJECT_SOURCE_DIR}/Source/Sandbox/GUI/*.cpp"
-                "${PROJECT_SOURCE_DIR}/Source/Sandbox/Game/*.cpp"
                 "${PROJECT_SOURCE_DIR}/Source/Sandbox/Core/*.cpp"
         )
 
         # Ogrid-Sandbox .H FILES
         file(GLOB_RECURSE Ogrid-Sandbox_H
-                "${PROJECT_SOURCE_DIR}/Source/Sandbox/GUI/*.h"
-                "${PROJECT_SOURCE_DIR}/Source/Sandbox/Game/*.h"
                 "${PROJECT_SOURCE_DIR}/Source/Sandbox/Core/*.h"
         )
