+++ Source/ogrid/GameDefinition/ConnectFourLogic.cpp
