+++ Source/ogrid/GameDefinition/ConnectFourTurnManager.cpp
