+++ Source/ogrid/GameDefinition/ConnectFourTurnManager.h
